Use int type for command line lengths.

+++ inc/common/cmd.h
@@ -175,8 +175,8 @@ char    *Cmd_ArgsFrom(int from);
 char    *Cmd_RawArgsFrom(int from);
 size_t  Cmd_ArgsBuffer(char *buffer, size_t size);
 size_t  Cmd_ArgvBuffer(int arg, char *buffer, size_t size);
-size_t  Cmd_ArgOffset(int arg);
-int     Cmd_FindArgForOffset(size_t offset);
+int     Cmd_ArgOffset(int arg);
+int     Cmd_FindArgForOffset(int offset);
 char    *Cmd_RawString(void);
 void    Cmd_Shift(void);
 // The functions that execute commands get their parameters with these
